requirements.txt
speedtest-cli==2.1.3
requests==2.31.0
psycopg2-binary==2.9.9
schedule==1.2.1
reportlab==4.1.0
python-dotenv==1.0.1
pytest==8.0.2
pytest-mock==3.12.0
pytest-env==1.1.3 